+++ packages/postgres/src/error.ts
@@ -1,7 +1,8 @@
 type VercelPostgresErrorCode =
   | 'invalid_connection_string'
   | 'missing_connection_string'
-  | 'invalid_connection_type';
+  | 'invalid_connection_type'
+  | 'incorrect_tagged_template_call';
 
 export class VercelPostgresError extends Error {
   public constructor(public code: VercelPostgresErrorCode, message: string) {