+++ packages/postgres/src/sql-template.ts
@@ -1,18 +1,32 @@
+import { VercelPostgresError } from './error';
+
 export type Primitive = string | number | boolean | undefined | null;
 
 export function sqlTemplate(
   strings: TemplateStringsArray,
   ...values: Primitive[]
 ): [string, Primitive[]] {
-  let result = '';
+  if (!isTemplateStringsArray(strings) || !Array.isArray(values)) {
+    throw new VercelPostgresError(
+      'incorrect_tagged_template_call',
+      "It looks like you tried to call `sql` as a function. Make sure to use it as a tagged template.\n\tExample: sql`SELECT * FROM users`, not sql('SELECT * FROM users')",
+    );
+  }
 
-  for (let i = 0; i < strings.length; i++) {
-    result += strings[i];
+  let result = strings[0] ?? '';
 
-    if (i < values.length) {
-      result += `$${i + 1}`;
-    }
+  for (let i = 1; i < strings.length; i++) {
+    result += `$${i}${strings[i] ?? ''}`;
   }
 
   return [result, values];
+}
+
+function isTemplateStringsArray(
+  strings: unknown,
+): strings is TemplateStringsArray {
+  return (
+    // @ts-expect-error - I don't know how to convince TS that an array can have a property
+    Array.isArray(strings) && 'raw' in strings && Array.isArray(strings.raw)
+  );
 }